move types inside src/utils/types
